🎨refactor: sms global로 변경
