+++ woorepie/src/main/java/com/piehouse/woorepie/global/service/SmsService.java
@@ -1,6 +1,10 @@
 package com.piehouse.woorepie.global.service;
 
+import com.piehouse.woorepie.global.dto.request.SmsAuthRequest;
+
 public interface SmsService {
+
+    void createSmsAuth(SmsAuthRequest smsAuthRequest);
 
     void sendSms(String toNumber, String code);
