+++ woorepie/src/main/java/com/piehouse/woorepie/global/service/impliment/SmsServiceImpl.java
@@ -1,5 +1,6 @@
 package com.piehouse.woorepie.global.service.impliment;
 
+import com.piehouse.woorepie.global.dto.request.SmsAuthRequest;
 import com.piehouse.woorepie.global.exception.CustomException;
 import com.piehouse.woorepie.global.exception.ErrorCode;
 import com.piehouse.woorepie.global.service.SmsService;
@@ -14,6 +15,7 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.stereotype.Service;
 
+import java.security.SecureRandom;
 import java.time.Duration;
 
 @Service
@@ -32,11 +34,28 @@ public class SmsServiceImpl implements SmsService {
     private final String REDIS_SMS_AUTH_KEY_PREFIX = "sms:auth:";
     private final RedisTemplate<String, String> redisStringTemplate;
     private DefaultMessageService messageService;
+    private final SecureRandom secureRandom = new SecureRandom();
 
     @PostConstruct
     public void init() {
         this.messageService = NurigoApp.INSTANCE
                 .initialize(apiKey, apiSecret, "https://api.coolsms.co.kr");
+    }
+
+
+    @Override
+    public void createSmsAuth(SmsAuthRequest smsAuthRequest) {
+        try {
+            String code = String.valueOf(secureRandom.nextInt(900_000) + 100_000);
+
+            // 인증번호 redis에 저장
+            redisSmsCode(smsAuthRequest.getPhoneNumber(), code);
+
+            // 문자 전송
+            sendSms(smsAuthRequest.getPhoneNumber(), code);
+        } catch (Exception e) {
+            throw new CustomException(ErrorCode.INTERNAL_ERROR);
+        }
     }
 
     // 인증 번호 문자 전송
